initial formatting and clean up for styling, moved useStyle config into its own file
