+++ src/components/LandingPage/LandingPage.jsx
@@ -1,56 +1,20 @@
 // First welcome page for new users, (if logged in, will reroute user to /discover)
 // reached at path /welcome
 
-import React, { useState } from 'react';
+import React from 'react';
 import { useHistory } from 'react-router-dom';
 
-import Image from '../Images/contriveLanding.png';
-
+import useStyles from './LandingPage.styles';
 
 // MATERIAL UI
-import { makeStyles } from '@material-ui/core/styles';
+
 import { Typography } from '@material-ui/core';
 import Box from '@material-ui/core/Box';
 import Button from '@material-ui/core/Button';
-import { spacing } from '@material-ui/system';
-
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    '& > *': {
-      margin: theme.spacing(1),
-    },
-  },
-  paperContainer: {
-    height: '625px',
-    // width:'100%',
-    // maxHeight: '100%',
-    // maxWidth: '100%',
-    backgroundImage: `url(${Image})`,
-    backgroundSize: 'cover',
-    backgroundRepeat: 'no-repeat',
-    backgroundPosition: 'center center',
-  },
-  box: {
-    height: 100,
-    display: "flex",
-    border: "1px solid black",
-    padding: 8
-  },
-  centerBox: {
-    justifyContent: "flex-center",
-    alignItems: "flex-center"
-  },
-  buttons: {
-    position: 'absolute',
-    bottom: theme.spacing.unit * 25,
-  }
-}));
 
 function LandingPage() {
   const history = useHistory();
   const classes = useStyles();
-  const [spacing, setSpacing] = React.useState(2);
 
   const onLogin = (event) => {
     history.push('/login');
@@ -67,16 +31,16 @@ function LandingPage() {
       <Box align="center" width="100%" className={classes.buttons}>
         <Box className={classes.root}>
 
-          <Button color="secondary" variant="contained" onClick={onLogin} style={{ height: 40 }}>
+          <Button color="secondary" variant="contained" onClick={onLogin} size="large">
             <Typography variant="body1">
               Login
-        </Typography>
+            </Typography>
           </Button>
 
-          <Button color="primary" variant="contained" onClick={onRegister} style={{ height: 40 }}>
+          <Button color="primary" variant="contained" onClick={onRegister} size="large">
             <Typography variant="body1">
               Register
-        </Typography>
+            </Typography>
           </Button>
 
         </Box>
